Trait concordium_contracts_common::Deserial

pub trait Deserial: Sized {
    fn deserial<R: Read>(_source: &mut R) -> ParseResult<Self>;
}

The Deserial trait provides a means of reading structures from byte-sources (Read).

Can be derived using #[derive(Deserial)] for most cases.

Required methods

fn deserial<R: Read>(_source: &mut R) -> ParseResult<Self>

Attempt to read a structure from a given source, failing if an error occurs during deserialization or reading.

Loading content...