Trait concordium_contracts_common::Write

pub trait Write {
    type Err: Default;
    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Err>;

    fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Err> { ... }

    fn write_u8(&mut self, x: u8) -> Result<(), Self::Err> { ... }

    fn write_u16(&mut self, x: u16) -> Result<(), Self::Err> { ... }

    fn write_u32(&mut self, x: u32) -> Result<(), Self::Err> { ... }

    fn write_u64(&mut self, x: u64) -> Result<(), Self::Err> { ... }

    fn write_i8(&mut self, x: i8) -> Result<(), Self::Err> { ... }

    fn write_i16(&mut self, x: i16) -> Result<(), Self::Err> { ... }

    fn write_i32(&mut self, x: i32) -> Result<(), Self::Err> { ... }

    fn write_i64(&mut self, x: i64) -> Result<(), Self::Err> { ... }
}

The Write trait provides functionality for writing to byte streams.

Associated Types

type Err: Default

Loading content...
Required methods

fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Err>

Try to write the given buffer into the output stream. If writes are successful returns the number of bytes written.

Loading content...
Provided methods

fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Err>

Attempt to write the entirety of the buffer to the output by repeatedly calling write until either no more output can written, or writing fails.

fn write_u8(&mut self, x: u8) -> Result<(), Self::Err>

Write a single byte to the output.

fn write_u16(&mut self, x: u16) -> Result<(), Self::Err>

Write a u16 in little endian.

fn write_u32(&mut self, x: u32) -> Result<(), Self::Err>

Write a u32 in little endian.

fn write_u64(&mut self, x: u64) -> Result<(), Self::Err>

Write a u64 in little endian.

fn write_i8(&mut self, x: i8) -> Result<(), Self::Err>

Write a i8 to the output.

fn write_i16(&mut self, x: i16) -> Result<(), Self::Err>

Write a i16 in little endian.

fn write_i32(&mut self, x: i32) -> Result<(), Self::Err>

Write a i32 in little endian.

fn write_i64(&mut self, x: i64) -> Result<(), Self::Err>

Write a i64 in little endian.

Loading content...