Type Definition concordium_contracts_common::OwnedPolicy

type OwnedPolicy = Policy<Vec<(AttributeTag, OwnedAttributeValue)>>;

A policy with a vector of attributes, fully allocated and owned. This is in contrast to a policy which is lazily read from a read source. The latter is useful for efficiency, this type is more useful for testing since the values are easier to construct.