Trait concordium_contracts_common::schema::SchemaType

pub trait SchemaType {
    fn get_type() -> Type;
}

The SchemaType trait provides means to generate a schema for structures. Schemas are used to make structures human readable and to avoid dealing directly with bytes, such as the contract state or parameters for contract interaction.

Can be derived using #[derive(SchemaType)] for most cases of structs and enums.

Required methods

fn get_type() -> Type