Trait concordium_contracts_common::Serial

pub trait Serial {
    fn serial<W: Write>(&self, _out: &mut W) -> Result<(), W::Err>;
}

The Serial trait provides a means of writing structures into byte-sinks (Write).

Can be derived using #[derive(Serial)] for most cases.

Required methods

fn serial<W: Write>(&self, _out: &mut W) -> Result<(), W::Err>

Attempt to write the structure into the provided writer, failing if only part of the structure could be written.

NB: We use Result instead of Option for better composability with other constructs.